requirements.txt
Flask>=2.2
pytesseract>=0.3.10
Pillow>=9.0
gunicorn>=20.1.0
opencv-python>=4.8.0
numpy>=1.24
